+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- New `create_wallet_if_missing` option to ease the creation of a designated bitcoind wallet (#76)
+
 ## 0.2.1 - 2021-01-14
 
 - Migrated to the [@bwt-dev](https://github.com/bwt-dev) github org and split up into: